backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
supabase==2.10.0
python-dotenv==1.0.1

psycopg2-binary==2.9.10
